+++ neoforge/src/main/java/com/lonkachu/stackable/Stackable.java
@@ -20,7 +20,6 @@ public class Stackable {
         // project.
 
         // Use NeoForge to bootstrap the Common mod.
-        Constants.LOG.info("Hello NeoForge world!");
         StackableMod.init();
 
         eventBus.addListener(this::ModifyDefaultComponentsEvent);